+++ lib/osf-components/addon/components/institutions-list/component.ts
@@ -1,0 +1,10 @@
+import Component from '@ember/component';
+import { localClassNames } from 'ember-osf-web/decorators/css-modules';
+import styles from './styles';
+import layout from './template';
+
+@localClassNames('InstitutionsList')
+export default class InstitutionsList extends Component {
+    layout = layout;
+    styles = styles;
+}