+++ mirage/serializers/node.ts
@@ -88,6 +88,16 @@ export default class NodeSerializer extends ApplicationSerializer<MirageNode> {
                     },
                 },
             },
+            affiliatedInstitutions: {
+                self: {
+                    href: `${apiUrl}/v2/nodes/${model.id}/relationships/institutions/`,
+                    meta: {},
+                },
+                related: {
+                    href: `${apiUrl}/v2/nodes/${model.id}/institutions/`,
+                    meta: this.buildRelatedLinkMeta(model, 'affiliatedInstitutions'),
+                },
+            },
         };
         if (model.attrs.parentId !== null) {
             const { parentId } = model.attrs;